zpprofile: Do not append script id to its body.
Since this was added, file names were changed to closely resemble original
script id. As the id alone is anyway already not enough to identify the
script (ex: when Skinnable magic is involved, multiple homonyms can exist
and be part of the same profiling result), so having it slightly extended
as file name should not matter much.