pprofile: Reorder cachegrind output.
Some callables are executed on the line they are defined, like dict or set
comprehensions. With original code, this would cause an "fn=" line to be
emitted for them, then their internal cost line, then call lines to them
without returning to parent function.
So instead do the rendering in 2 passes so caller information is not lost.
Also, make local naming a bit more coherent.